Moving GC to the main thread. See the message below
somehow this is the key to move the performance
form ~100 Gbps to 150+ Gbps on 16 cores I don't
know running GC here has a different effect than
running in the start and end of ExecuteTest framework.
Also, here it runs synchronously.

For now the problem is gone with large enough young
generation and G1GC - if we get time we  might return
to it.